fix(813): Handle nullable value classes
Fix https://github.com/mockk/mockk/issues/813

Problem:
When trying to mock a function that returns a nullable value class, a ClassCastException was thrown. Thats because  mockk is using the boxedClass.
Nullable value classes should be handled differently because they are not inlined.

Implementation:
When the return type of the function or a member property is nullable do not use the boxedClass but use the value class itself

fix